[A|Weather@feat/weather] Assets added + activity started

+++ app/src/main/java/com/riders/testing/model/weather/Coordinates.java
@@ -5,24 +5,24 @@ import com.google.gson.annotations.SerializedName;
 public class Coordinates {
 
     @SerializedName("lon")
-    int longitude;
+    double longitude;
 
     @SerializedName("lat")
-    int latitude;
+    double latitude;
 
-    public int getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(int longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 
-    public int getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(int latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 }
